--- src/middlewared/middlewared/plugins/alert/alert.py
@@ -1,1108 +1,0 @@
-from __future__ import annotations
-
-from collections import defaultdict
-import copy
-from dataclasses import dataclass
-from datetime import datetime, timezone
-import errno
-from itertools import zip_longest
-import textwrap
-import time
-import typing
-from typing import Any
-import uuid
-
-import html2text
-from truenas_api_client.exc import ReserveFDException
-
-from middlewared.alert.base import (
-    Alert,
-    AlertCategory,
-    AlertClass,
-    AlertClassConfig,
-    AlertLevel,
-    AlertSource,
-    DismissableAlertClass,
-    OneShotAlertClass,
-    UnavailableException,
-    alert_category_names,
-)
-from middlewared.alert.base import (
-    AlertService as _AlertService,
-)
-import middlewared.alert.source  # noqa: F401
-from middlewared.api import Event, api_method
-from middlewared.api.base import BaseModel
-from middlewared.api.current import (
-    AlertDismissArgs,
-    AlertDismissResult,
-    AlertListAddedEvent,
-    AlertListArgs,
-    AlertListCategoriesArgs,
-    AlertListCategoriesResult,
-    AlertListChangedEvent,
-    AlertListPoliciesArgs,
-    AlertListPoliciesResult,
-    AlertListRemovedEvent,
-    AlertListResult,
-    AlertRestoreArgs,
-    AlertRestoreResult,
-)
-from middlewared.plugins.failover_.remote import NETWORK_ERRORS
-from middlewared.service import Service, job, periodic, private
-from middlewared.service_exception import CallError, NetworkActivityDisabled
-import middlewared.sqlalchemy as sa
-from middlewared.utils.time_utils import utc_now
-
-if typing.TYPE_CHECKING:
-    from collections.abc import Callable, Iterable
-
-    from middlewared.main import Middleware
-
-POLICIES: list[str] = ["IMMEDIATELY", "HOURLY", "DAILY", "NEVER"]
-DEFAULT_POLICY: str = "IMMEDIATELY"
-SEND_ALERTS_ON_READY: bool = False
-# The below value come from observation from support of how long an M-series boot can take.
-FAILOVER_ALERTS_BACKOFF_SECS: int = 900
-
-
-@dataclass(slots=True, frozen=True, kw_only=True)
-class AlertFailoverInfo:
-    this_node: str
-    other_node: str
-    run_on_backup_node: bool
-    run_failover_related: bool
-
-
-@dataclass(slots=True, frozen=True)
-class AlertSourceLock:
-    source_name: str
-    expires_at: float
-
-
-class AlertModel(sa.Model):
-    __tablename__ = 'system_alert'
-
-    id = sa.Column(sa.Integer(), primary_key=True)
-    node = sa.Column(sa.String(100))
-    source = sa.Column(sa.Text())
-    key = sa.Column(sa.Text())
-    datetime = sa.Column(sa.DateTime())
-    last_occurrence = sa.Column(sa.DateTime())
-    text = sa.Column(sa.Text())
-    args = sa.Column(sa.JSON(None))
-    dismissed = sa.Column(sa.Boolean())
-    uuid = sa.Column(sa.Text())
-    klass = sa.Column(sa.Text())
-
-
-@dataclass(kw_only=True)
-class AlertSourceRunFailedAlert(AlertClass):
-    config = AlertClassConfig(
-        category=AlertCategory.SYSTEM,
-        level=AlertLevel.CRITICAL,
-        title="Alert Check Failed",
-        text="Failed to check for alert %(source_name)s: %(traceback)s",
-        exclude_from_list=True,
-    )
-    source_name: str
-    traceback: str
-
-
-@dataclass(kw_only=True)
-class AlertSourceRunFailedOnBackupNodeAlert(AlertClass):
-    config = AlertClassConfig(
-        category=AlertCategory.SYSTEM,
-        level=AlertLevel.CRITICAL,
-        title="Alert Check Failed (Standby Controller)",
-        text="Failed to check for alert %(source_name)s on standby controller: %(traceback)s",
-        exclude_from_list=True,
-    )
-    source_name: str
-    traceback: str
-
-
-@dataclass(kw_only=True)
-class AutomaticAlertFailedAlert(OneShotAlertClass):
-    config = AlertClassConfig(
-        category=AlertCategory.SYSTEM,
-        level=AlertLevel.WARNING,
-        title="Failed to Notify TrueNAS About Alert",
-        text=textwrap.dedent("""\
-            Creating an automatic alert for TrueNAS about system %(serial)s failed: %(error)s.
-            Please contact TrueNAS Support: https://www.truenas.com/support/
-
-            Alert:
-
-            %(alert)s
-        """),
-        exclude_from_list=True,
-        deleted_automatically=False,
-    )
-    serial: str
-    error: str
-    alert: str
-
-
-class AlertPolicy:
-    def __init__(self, key: Callable[[datetime], Any] = lambda now: now) -> None:
-        self.key = key
-        self.last_key_value: Any = None
-        self.last_key_value_alerts: dict[str, Alert[Any]] = {}
-
-    def receive_alerts(self, now: datetime, alerts: list[Alert[Any]]) -> tuple[list[Alert[Any]], list[Alert[Any]]]:
-        alerts_by_uuid = {alert.uuid: alert for alert in alerts}
-        gone_alerts: list[Alert[Any]] = []
-        new_alerts: list[Alert[Any]] = []
-        key = self.key(now)
-        if key != self.last_key_value:
-            gone_alerts = [alert for alert in self.last_key_value_alerts.values() if alert.uuid not in alerts_by_uuid]
-            new_alerts = [alert for alert in alerts_by_uuid.values() if alert.uuid not in self.last_key_value_alerts]
-
-            self.last_key_value = key
-            self.last_key_value_alerts = alerts_by_uuid
-
-        return gone_alerts, new_alerts
-
-    def delete_alert(self, alert: Alert[Any]) -> None:
-        self.last_key_value_alerts.pop(alert.uuid, None)
-
-
-def get_alert_level(alert: Alert[Any], classes: dict[str, Any]) -> AlertLevel:
-    return AlertLevel[classes.get(alert.instance.config.name, {}).get("level", alert.instance.config.level.name)]
-
-
-def get_alert_policy(alert: Alert[Any], classes: dict[str, Any]) -> str:
-    return classes.get(alert.instance.config.name, {}).get("policy", DEFAULT_POLICY)  # type: ignore[no-any-return]
-
-
-def partition[T](predicate: Callable[[T], Any], iterable: Iterable[T]) -> tuple[list[T], list[T]]:
-    """Split *iterable* into ``(matching, non_matching)`` lists by *predicate*.
-
-    Order within each list is preserved from the input. The predicate is
-    evaluated exactly once per item.
-    """
-    matching: list[T] = []
-    non_matching: list[T] = []
-    for item in iterable:
-        if predicate(item):
-            matching.append(item)
-        else:
-            non_matching.append(item)
-    return matching, non_matching
-
-
-class AlertSerializer:
-    def __init__(self, middleware: Middleware) -> None:
-        self.middleware = middleware
-
-        self.initialized: bool = False
-        self.product_type: str = ""
-        self.classes: dict[str, dict[str, Any]] = {}
-        self.nodes: dict[str, str] = {}
-
-    async def serialize(self, alert: Alert[Any]) -> dict[str, Any]:
-        await self._ensure_initialized()
-
-        return dict(
-            id=alert.uuid,
-            uuid=alert.uuid,
-            source=alert.source,
-            klass=alert.instance.config.name,
-            args=alert.instance.args(),
-            node=self.nodes[alert.node],
-            key=alert.key,
-            datetime=alert.datetime,
-            last_occurrence=alert.last_occurrence,
-            dismissed=alert.dismissed,
-            mail=alert.mail,
-            text=alert.text,
-            level=self.classes.get(alert.instance.config.name, {}).get("level", alert.instance.config.level.name),
-            formatted=alert.formatted,
-            one_shot=isinstance(alert.instance, OneShotAlertClass) and not alert.instance.config.deleted_automatically
-        )
-
-    async def get_alert_class(self, alert: Alert[Any]) -> dict[str, Any]:
-        await self._ensure_initialized()
-        return self.classes.get(alert.instance.config.name, {})
-
-    async def should_show_alert(self, alert: Alert[Any]) -> bool:
-        await self._ensure_initialized()
-
-        if self.product_type not in alert.instance.config.products:
-            return False
-
-        if (await self.get_alert_class(alert)).get("policy") == "NEVER":
-            return False
-
-        return True
-
-    async def _ensure_initialized(self) -> None:
-        if not self.initialized:
-            self.product_type = await self.middleware.call2(self.middleware.services.alert.product_type)
-            self.classes = (
-                await self.middleware.call2(self.middleware.services.alertclasses.config)
-            ).classes  # type: ignore[assignment]
-            self.nodes = await self.middleware.call2(self.middleware.services.alert.node_map)
-
-            self.initialized = True
-
-
-class AlertOneshotDeleteArgs(BaseModel):
-    klass: str | list[str]
-    query: Any = None
-
-
-class AlertOneshotDeleteResult(BaseModel):
-    result: None
-
-
-List = list  # avoid shadowing by AlertService.list
-
-
-class AlertService(Service):
-    alert_sources_errors: set[str] = set()
-
-    class Config:
-        cli_namespace = "system.alert"
-        events = [
-            Event(
-                name="alert.list",
-                description="Sent on alert changes.",
-                roles=["ALERT_LIST_READ"],
-                models={
-                    "ADDED": AlertListAddedEvent,
-                    "CHANGED": AlertListChangedEvent,
-                    "REMOVED": AlertListRemovedEvent,
-                },
-            ),
-        ]
-
-    def __init__(self, middleware: Middleware) -> None:
-        super().__init__(middleware)
-
-        self.alert_sources: dict[str, AlertSource] = {
-            name: cls(middleware) for name, cls in AlertSource.by_name.items()
-        }
-        self.blocked_sources: defaultdict[str, set[str]] = defaultdict(set)
-        self.sources_locks: dict[str, AlertSourceLock] = {}
-
-        self.blocked_failover_alerts_until: float = 0
-
-        self.sources_run_times: defaultdict[str, dict[str, Any]] = defaultdict(lambda: {
-            "last": [],
-            "max": 0,
-            "total_count": 0,
-            "total_time": 0,
-        })
-
-        self.node: str = "A"
-        self.alerts: list[Alert[Any]] = []
-        self.alert_source_last_run: defaultdict[str, datetime] = defaultdict(lambda: datetime.min)
-        self.policies: dict[str, AlertPolicy] = {}
-
-    @private
-    async def initialize(self, load: bool = True) -> None:
-        is_enterprise: bool = await self.middleware.call("system.is_enterprise")
-
-        self.node = "A"
-        if is_enterprise:
-            if await self.middleware.call("failover.node") == "B":
-                self.node = "B"
-
-        self.alerts = []
-        if load:
-            alerts_uuids: set[str] = set()
-            alerts_by_classes: defaultdict[str, list[Alert[Any]]] = defaultdict(list)
-            for alert in await self.middleware.call("datastore.query", "system.alert"):
-                del alert["id"]
-
-                if alert["source"] and alert["source"] not in self.alert_sources:
-                    self.logger.info("Alert source %r is no longer present", alert["source"])
-                    continue
-
-                class_name: str = alert.pop("klass")
-                try:
-                    klass = AlertClass.by_name[class_name]
-                except KeyError:
-                    self.logger.info("Alert class %r is no longer present", class_name)
-                    continue
-
-                alert["_uuid"] = alert.pop("uuid")
-                alert["_source"] = alert.pop("source")
-                alert["_key"] = alert.pop("key")
-                alert["_text"] = alert.pop("text")
-
-                args = alert.pop("args")
-                try:
-                    instance = klass.from_args(args)
-                    alert = Alert(instance, **alert)
-                except Exception as e:
-                    self.logger.info("Error loading alert class %r: %r", class_name, e)
-                    continue
-
-                if alert.uuid not in alerts_uuids:
-                    alerts_uuids.add(alert.uuid)
-                    alerts_by_classes[alert.instance.config.name].append(alert)
-
-            for alerts in alerts_by_classes.values():
-                if isinstance(alerts[0].instance, OneShotAlertClass):
-                    try:
-                        alerts = await alerts[0].instance.load(self.middleware, alerts)
-                    except Exception as e:
-                        self.logger.info("Error loading one-shot alert %r: %r", alerts[0].instance, e)
-                        continue
-
-                self.alerts.extend(alerts)
-        else:
-            await self.call2(self.s.alert.flush_alerts)
-
-        self.alert_source_last_run = defaultdict(lambda: datetime.min)
-
-        self.policies = {
-            "IMMEDIATELY": AlertPolicy(),
-            "HOURLY": AlertPolicy(lambda d: (d.date(), d.hour)),
-            "DAILY": AlertPolicy(lambda d: (d.date())),
-            "NEVER": AlertPolicy(lambda d: None),
-        }
-        for policy in self.policies.values():
-            policy.receive_alerts(utc_now(), self.alerts)
-
-    @private
-    async def terminate(self) -> None:
-        await self.call2(self.s.alert.flush_alerts)
-
-    @api_method(AlertListPoliciesArgs, AlertListPoliciesResult, roles=['ALERT_LIST_READ'])
-    async def list_policies(self) -> list[str]:
-        """
-        List all alert policies which indicate the frequency of the alerts.
-        """
-        return POLICIES
-
-    @api_method(AlertListCategoriesArgs, AlertListCategoriesResult, roles=['ALERT_LIST_READ'])
-    async def list_categories(self, options: dict[str, Any]) -> list[dict[str, Any]]:
-        """
-        List all types of alerts which the system can issue.
-        """
-
-        product_type: str = await self.call2(self.s.alert.product_type)
-
-        classes: list[type[AlertClass]] = []
-        for alert_class in AlertClass.classes:
-            if not (options["include_all_products"] or product_type in alert_class.config.products):
-                continue
-
-            if not (options["include_hidden_classes"] or not alert_class.config.exclude_from_list):
-                continue
-
-            classes.append(alert_class)
-
-        return [
-            {
-                "id": alert_category.name,
-                "title": alert_category_names[alert_category],
-                "classes": sorted(
-                    [
-                        {
-                            "id": alert_class.config.name,
-                            "title": alert_class.config.title,
-                            "level": alert_class.config.level.name,
-                            "product_types": list(alert_class.config.products),
-                            "proactive_support": alert_class.config.proactive_support,
-                        }
-                        for alert_class in classes
-                        if alert_class.config.category == alert_category
-                    ],
-                    key=lambda klass: klass["title"]
-                )
-            }
-            for alert_category in AlertCategory
-            if any(alert_class.config.category == alert_category for alert_class in classes)
-        ]
-
-    @api_method(AlertListArgs, AlertListResult, roles=['ALERT_LIST_READ'])
-    async def list(self) -> list[dict[str, Any]]:
-        """
-        List all types of alerts including active/dismissed currently in the system.
-        """
-
-        as_ = AlertSerializer(self.middleware)
-        classes = (await self.call2(self.s.alertclasses.config)).classes
-
-        return [
-            await as_.serialize(alert)
-            for alert in sorted(
-                self.alerts,
-                key=lambda alert: (
-                    -get_alert_level(alert, classes).value,
-                    alert.instance.config.title,
-                    alert.datetime,
-                ),
-            )
-            if await as_.should_show_alert(alert)
-        ]
-
-    @private
-    async def node_map(self) -> dict[str, str]:
-        nodes: dict[str, str] = {
-            'A': 'Controller A',
-            'B': 'Controller B',
-        }
-        if await self.middleware.call('failover.licensed'):
-            node: str = await self.middleware.call('failover.node')
-            status: str = await self.middleware.call('failover.status')
-            if status == 'MASTER':
-                if node == 'A':
-                    nodes = {
-                        'A': 'Active Controller (A)',
-                        'B': 'Standby Controller (B)',
-                    }
-                else:
-                    nodes = {
-                        'A': 'Standby Controller (A)',
-                        'B': 'Active Controller (B)',
-                    }
-            else:
-                nodes[node] = f'{status.title()} Controller ({node})'
-
-        return nodes
-
-    def __alert_by_uuid(self, uuid: str) -> Alert[Any] | None:
-        try:
-            return [a for a in self.alerts if a.uuid == uuid][0]
-        except IndexError:
-            return None
-
-    @api_method(AlertDismissArgs, AlertDismissResult, roles=['ALERT_LIST_WRITE'])
-    async def dismiss(self, uuid: str) -> None:
-        """
-        Dismiss `id` alert.
-        """
-
-        alert = self.__alert_by_uuid(uuid)
-        if alert is None:
-            return
-
-        if isinstance(alert.instance, DismissableAlertClass):
-            related_alerts, unrelated_alerts = partition(
-                lambda a: (a.node, a.instance.config.name) == (alert.node, alert.instance.config.name),
-                self.alerts,
-            )
-            left_alerts = await alert.instance.dismiss(self.middleware, related_alerts, alert)
-            for deleted_alert in related_alerts:
-                if deleted_alert not in left_alerts:
-                    self._delete_on_dismiss(deleted_alert)
-        elif isinstance(alert.instance, OneShotAlertClass) and not alert.instance.config.deleted_automatically:
-            self._delete_on_dismiss(alert)
-        else:
-            alert.dismissed = True
-            await self._send_alert_changed_event(alert)
-
-    def _delete_on_dismiss(self, alert: Alert[Any]) -> None:
-        try:
-            self.alerts.remove(alert)
-            removed = True
-        except ValueError:
-            removed = False
-
-        for policy in self.policies.values():
-            policy.delete_alert(alert)
-
-        if removed:
-            self._send_alert_deleted_event(alert)
-
-    @api_method(AlertRestoreArgs, AlertRestoreResult, roles=['ALERT_LIST_WRITE'])
-    async def restore(self, uuid: str) -> None:
-        """
-        Restore `id` alert which had been dismissed.
-        """
-
-        alert = self.__alert_by_uuid(uuid)
-        if alert is None:
-            return
-
-        alert.dismissed = False
-
-        await self._send_alert_changed_event(alert)
-
-    async def _send_alert_changed_event(self, alert: Alert[Any]) -> None:
-        as_ = AlertSerializer(self.middleware)
-        if await as_.should_show_alert(alert):
-            self.middleware.send_event("alert.list", "CHANGED", id=alert.uuid, fields=await as_.serialize(alert))
-
-    def _send_alert_deleted_event(self, alert: Alert[Any]) -> None:
-        self.middleware.send_event("alert.list", "REMOVED", id=alert.uuid)
-
-    @periodic(60)
-    @private
-    @job(lock="process_alerts", transient=True, lock_queue_size=1)
-    async def process_alerts(self, job: Any) -> None:
-        if not await self.__should_run_or_send_alerts():
-            return
-
-        valid_alerts = copy.deepcopy(self.alerts)
-        await self.__run_alerts()
-
-        self.__expire_alerts()
-
-        if not await self.__should_run_or_send_alerts():
-            self.alerts = valid_alerts
-            return
-
-        await self.call2(self.s.alert.send_alerts)
-
-    @private
-    @job(lock="process_alerts", transient=True)
-    async def send_alerts(self, job: Any) -> None:
-        global SEND_ALERTS_ON_READY
-
-        if await self.middleware.call("system.state") != "READY":
-            SEND_ALERTS_ON_READY = True
-            return
-
-        product_type: str = await self.call2(self.s.alert.product_type)
-        classes: dict[str, Any] = (await self.call2(self.s.alertclasses.config)).classes
-
-        now = utc_now()
-        for policy_name, policy in self.policies.items():
-            gone_alerts, new_alerts = policy.receive_alerts(now, self.alerts)
-
-            for alert_service_desc in await self.call2(self.s.alertservice.query, [["enabled", "=", True]]):
-                service_level = AlertLevel[alert_service_desc.level]
-
-                service_alerts = [
-                    alert for alert in self.alerts
-                    if (
-                        product_type in alert.instance.config.products and
-                        get_alert_level(alert, classes).value >= service_level.value and
-                        get_alert_policy(alert, classes) != "NEVER"
-                    )
-                ]
-                service_gone_alerts = [
-                    alert for alert in gone_alerts
-                    if (
-                        product_type in alert.instance.config.products and
-                        get_alert_level(alert, classes).value >= service_level.value and
-                        get_alert_policy(alert, classes) == policy_name
-                    )
-                ]
-                service_new_alerts = [
-                    alert for alert in new_alerts
-                    if (
-                        product_type in alert.instance.config.products and
-                        get_alert_level(alert, classes).value >= service_level.value and
-                        get_alert_policy(alert, classes) == policy_name
-                    )
-                ]
-                for gone_alert in list(service_gone_alerts):
-                    for new_alert in service_new_alerts:
-                        if (
-                            gone_alert.instance.config.name == new_alert.instance.config.name and
-                            gone_alert.key == new_alert.key
-                        ):
-                            service_gone_alerts.remove(gone_alert)
-                            service_new_alerts.remove(new_alert)
-                            break
-
-                if not service_gone_alerts and not service_new_alerts:
-                    continue
-
-                factory = _AlertService.by_name[alert_service_desc.attributes.type]
-                alert_service = factory(
-                    self.middleware,
-                    alert_service_desc.attributes.model_dump(context={"expose_secrets": True}),
-                )
-
-                alerts = [alert for alert in service_alerts if not alert.dismissed]
-                service_gone_alerts = [alert for alert in service_gone_alerts if not alert.dismissed]
-                service_new_alerts = [alert for alert in service_new_alerts if not alert.dismissed]
-
-                if alerts or service_gone_alerts or service_new_alerts:
-                    try:
-                        await alert_service.send(alerts, service_gone_alerts, service_new_alerts)
-                    except Exception:
-                        self.logger.error(
-                            "Error in alert service %r", alert_service_desc.attributes.type, exc_info=True,
-                        )
-
-            if policy_name == "IMMEDIATELY":
-                as_ = AlertSerializer(self.middleware)
-                for alert in gone_alerts:
-                    if await as_.should_show_alert(alert):
-                        self._send_alert_deleted_event(alert)
-                for alert in new_alerts:
-                    if await as_.should_show_alert(alert):
-                        self.middleware.send_event(
-                            "alert.list", "ADDED", id=alert.uuid, fields=await as_.serialize(alert),
-                        )
-
-                for alert in new_alerts:
-                    if alert.mail:
-                        try:
-                            await self.call2(self.s.mail.send, alert.mail)
-                        except NetworkActivityDisabled:
-                            pass
-
-                if await self.middleware.call("system.is_enterprise"):
-                    gone_proactive_support_alerts = [
-                        alert
-                        for alert in gone_alerts
-                        if (
-                            alert.instance.config.proactive_support and
-                            (await as_.get_alert_class(alert)).get("proactive_support", True) and
-                            alert.instance.config.proactive_support_notify_gone
-                        )
-                    ]
-                    new_proactive_support_alerts = [
-                        alert
-                        for alert in new_alerts
-                        if (
-                            alert.instance.config.proactive_support and
-                            (await as_.get_alert_class(alert)).get("proactive_support", True)
-                        )
-                    ]
-                    if gone_proactive_support_alerts or new_proactive_support_alerts:
-                        if await self.middleware.call("support.is_available_and_enabled"):
-                            support = await self.middleware.call("support.config")
-
-                            msg: list[str] = []
-                            if gone_proactive_support_alerts:
-                                msg.append("The following alerts were cleared:")
-                                msg += [f"* {html2text.html2text(alert.formatted)}"
-                                        for alert in gone_proactive_support_alerts]
-                            if new_proactive_support_alerts:
-                                msg.append("The following new alerts appeared:")
-                                msg += [f"* {html2text.html2text(alert.formatted)}"
-                                        for alert in new_proactive_support_alerts]
-
-                            serial: str = (
-                                await self.middleware.call("system.dmidecode_info")
-                            )["system-serial-number"]
-
-                            for name, verbose_name in await self.middleware.call("support.fields"):
-                                value = support[name]
-                                if value:
-                                    msg += ["", "{}: {}".format(verbose_name, value)]
-
-                            msg_str: str = "\n".join(msg)
-
-                            job = await self.middleware.call("support.new_ticket", {
-                                "title": "Automatic alert (%s)" % serial,
-                                "body": msg_str,
-                                "attach_debug": False,
-                                "category": "Hardware",
-                                "criticality": "Loss of Functionality",
-                                "environment": "Production",
-                                "name": "Automatic Alert",
-                                "email": "auto-support@truenas.com",
-                                "phone": "-",
-                            })
-                            await job.wait()
-                            if job.error:
-                                await self.call2(self.s.alert.oneshot_create, AutomaticAlertFailedAlert(
-                                    serial=serial, alert=msg_str, error=str(job.error)))
-
-    def __uuid(self) -> str:
-        return str(uuid.uuid4())
-
-    async def __should_run_or_send_alerts(self) -> bool:
-        if await self.middleware.call('system.state') != 'READY':
-            return False
-
-        if await self.middleware.call('failover.licensed'):
-            status: str = await self.middleware.call('failover.status')
-            if status == 'BACKUP' or await self.middleware.call('failover.in_progress'):
-                return False
-
-        return True
-
-    async def __get_failover_info(self) -> AlertFailoverInfo:
-        this_node, other_node = "A", "B"
-        run_on_backup_node = False
-        run_failover_related = await self.middleware.call("failover.licensed")
-        if run_failover_related:
-            if await self.middleware.call("failover.node") != "A":
-                this_node, other_node = "B", "A"
-
-            run_failover_related = time.monotonic() > self.blocked_failover_alerts_until
-            if run_failover_related:
-                args = ([], {"connect_timeout": 2})  # type: ignore[var-annotated]
-
-                # Do not run on backup if there is a software version mismatch
-                try:
-                    rem_ver = await self.middleware.call("failover.call_remote", "system.version", *args)
-                    run_on_backup_node = (await self.middleware.call("system.version")) == rem_ver
-                except Exception:
-                    pass
-
-                # Do not run on backup if the other node is not READY
-                if run_on_backup_node:
-                    try:
-                        run_on_backup_node = (await self.middleware.call(
-                            "failover.call_remote", "system.state", *args
-                        )) == "READY"
-                    except Exception:
-                        pass
-
-                # Do not run on backup if the other node is not BACKUP
-                if run_on_backup_node:
-                    try:
-                        run_on_backup_node = (await self.middleware.call(
-                            "failover.call_remote", "system.status", *args
-                        )) == "BACKUP"
-                    except Exception:
-                        pass
-
-                if run_on_backup_node:
-                    # If BACKUP node is in good shape, check whether it
-                    # has been booted for long enough.
-                    try:
-                        run_on_backup_node = (await self.middleware.call(
-                            "failover.call_remote", "system.time_info", *args
-                        ))['uptime_seconds'] > FAILOVER_ALERTS_BACKOFF_SECS
-                    except Exception:
-                        pass
-
-        return AlertFailoverInfo(
-            this_node=this_node,
-            other_node=other_node,
-            run_on_backup_node=run_on_backup_node,
-            run_failover_related=run_failover_related
-        )
-
-    async def __handle_locked_alert_source(
-        self, name: str, this_node: str, other_node: str,
-    ) -> tuple[List[Alert[Any]], List[Alert[Any]], set[str]]:
-        this_node_alerts: list[Alert[Any]] = []
-        other_node_alerts: list[Alert[Any]] = []
-        locked = self.blocked_sources[name]
-        if locked:
-            self.logger.debug("Not running alert source %r because it is blocked", name)
-            for i in filter(lambda x: x.source == name, self.alerts):
-                if i.node == this_node:
-                    this_node_alerts.append(i)
-                elif i.node == other_node:
-                    other_node_alerts.append(i)
-        return this_node_alerts, other_node_alerts, locked
-
-    async def __run_other_node_alert_source(self, name: str) -> List[Alert[Any]]:
-        keys = ("datetime", "last_occurrence", "dismissed", "mail")
-        other_node_alerts: list[Alert[Any]] = []
-        try:
-            try:
-                for alert in await self.middleware.call("failover.call_remote", "alert.run_source", [name]):
-                    klass = AlertClass.by_name[alert["klass"]]
-                    instance = klass.from_args(alert["args"])
-                    other_node_alerts.append(
-                        Alert(instance,
-                              **{k: v for k, v in alert.items() if k in keys},
-                              _source=alert["source"],
-                              _key=alert["key"])
-                    )
-            except CallError as e:
-                if e.errno not in NETWORK_ERRORS + (CallError.EALERTCHECKERUNAVAILABLE,):
-                    raise
-        except ReserveFDException:
-            self.logger.debug('Failed to reserve a privileged port')
-        except Exception as e:
-            other_node_alerts = [Alert(
-                AlertSourceRunFailedOnBackupNodeAlert(source_name=name, traceback=str(e)),
-                _source=name
-            )]
-
-        return other_node_alerts
-
-    async def __run_alerts(self) -> None:
-        product_type: str = await self.call2(self.s.alert.product_type)
-        fi = await self.__get_failover_info()
-        for k, source_lock in list(self.sources_locks.items()):
-            if source_lock.expires_at <= time.monotonic():
-                await self.call2(self.s.alert.unblock_source, k)
-
-        for alert_source in self.alert_sources.values():
-            if product_type not in alert_source.products:
-                continue
-
-            if alert_source.failover_related and not fi.run_failover_related:
-                continue
-
-            if alert_source.require_stable_peer and not fi.run_on_backup_node:
-                continue
-
-            if not alert_source.schedule.should_run(utc_now(), self.alert_source_last_run[alert_source.name]):
-                continue
-
-            self.alert_source_last_run[alert_source.name] = utc_now()
-
-            this_node_alerts, other_node_alerts, locked = await self.__handle_locked_alert_source(
-                alert_source.name, fi.this_node, fi.other_node
-            )
-            if not locked:
-                self.logger.trace("Running alert source: %r", alert_source.name)
-                try:
-                    this_node_alerts = await self.__run_source(alert_source.name)
-                except UnavailableException:
-                    pass
-
-                if fi.run_on_backup_node and alert_source.run_on_backup_node:
-                    other_node_alerts = await self.__run_other_node_alert_source(alert_source.name)
-
-            for talert, oalert in zip_longest(this_node_alerts, other_node_alerts, fillvalue=None):
-                if talert is not None:
-                    talert.node = fi.this_node
-                    self.__handle_alert(talert)
-                if oalert is not None:
-                    oalert.node = fi.other_node
-                    self.__handle_alert(oalert)
-
-            self.alerts = (
-                [a for a in self.alerts if a.source != alert_source.name] + this_node_alerts + other_node_alerts
-            )
-
-    def __handle_alert(self, alert: Alert[Any]) -> None:
-        try:
-            existing_alert = [
-                a for a in self.alerts
-                if (a.node, a.source, a.instance.config.name, a.key) == (
-                    alert.node, alert.source, alert.instance.config.name, alert.key
-                )
-            ][0]
-        except IndexError:
-            existing_alert = None
-
-        if existing_alert is None:
-            alert.uuid = self.__uuid()
-        else:
-            alert.uuid = existing_alert.uuid
-        if existing_alert is None:
-            alert.datetime = alert.datetime or utc_now()
-            if alert.datetime.tzinfo is not None:
-                alert.datetime = alert.datetime.astimezone(timezone.utc).replace(tzinfo=None)
-        else:
-            alert.datetime = existing_alert.datetime
-        alert.last_occurrence = utc_now()
-        if existing_alert is None:
-            alert.dismissed = False
-        else:
-            alert.dismissed = existing_alert.dismissed
-
-    def __expire_alerts(self) -> None:
-        self.alerts = list(filter(lambda alert: not self.__should_expire_alert(alert), self.alerts))
-
-    def __should_expire_alert(self, alert: Alert[Any]) -> bool:
-        if isinstance(alert.instance, OneShotAlertClass):
-            if alert.instance.config.expires_after is not None:
-                return alert.last_occurrence < utc_now() - alert.instance.config.expires_after  # type: ignore[operator]
-
-        return False
-
-    @private
-    async def sources_stats(self) -> dict[str, Any]:
-        return {
-            k: {"avg": v["total_time"] / v["total_count"] if v["total_count"] != 0 else 0, **v}
-            for k, v in sorted(self.sources_run_times.items(), key=lambda t: t[0])
-        }
-
-    @private
-    async def run_source(self, source_name: str) -> List[dict[str, Any]]:
-        try:
-            return [dict(alert.__dict__, instance=None, klass=alert.instance.config.name, args=alert.instance.args())
-                    for alert in await self.__run_source(source_name)]
-        except UnavailableException:
-            raise CallError("This alert checker is unavailable", CallError.EALERTCHECKERUNAVAILABLE)
-
-    @private
-    async def block_source(self, source_name: str, timeout: int = 3600) -> str:
-        if source_name not in self.alert_sources:
-            raise CallError("Invalid alert source")
-
-        lock = str(uuid.uuid4())
-        self.blocked_sources[source_name].add(lock)
-        self.sources_locks[lock] = AlertSourceLock(source_name, time.monotonic() + timeout)
-        return lock
-
-    @private
-    async def unblock_source(self, lock: str) -> None:
-        source_lock = self.sources_locks.pop(lock, None)
-        if source_lock:
-            self.blocked_sources[source_lock.source_name].remove(lock)
-
-    @private
-    async def block_failover_alerts(self) -> None:
-        self.blocked_failover_alerts_until = time.monotonic() + FAILOVER_ALERTS_BACKOFF_SECS
-
-    async def __run_source(self, source_name: str) -> List[Alert[Any]]:
-        alert_source = self.alert_sources[source_name]
-
-        start = time.monotonic()
-        try:
-            alerts = (await alert_source.check()) or []
-        except UnavailableException:
-            raise
-        except Exception as e:
-            if source_name not in self.alert_sources_errors:
-                self.logger.error("Error checking for alert %r", alert_source.name, exc_info=True)
-                self.alert_sources_errors.add(source_name)
-
-            alerts = [
-                Alert(AlertSourceRunFailedAlert(
-                    source_name=alert_source.name,
-                    traceback=str(e),
-                ))
-            ]
-        else:
-            self.alert_sources_errors.discard(source_name)
-            if not isinstance(alerts, list):
-                alerts = [alerts]
-        finally:
-            run_time = time.monotonic() - start
-            source_stat = self.sources_run_times[source_name]
-            source_stat["last"] = source_stat["last"][-9:] + [run_time]
-            source_stat["max"] = max(source_stat["max"], run_time)
-            source_stat["total_count"] += 1
-            source_stat["total_time"] += run_time
-
-        keys: set[str] = set()
-        unique_alerts: list[Alert[Any]] = []
-        for alert in alerts:
-            if alert.key in keys:
-                continue
-
-            keys.add(alert.key)
-            unique_alerts.append(alert)
-        alerts = unique_alerts
-
-        for alert in alerts:
-            alert.source = source_name
-
-        return alerts
-
-    @periodic(3600, run_on_start=False)
-    @private
-    async def flush_alerts(self) -> None:
-        if await self.middleware.call('failover.licensed'):
-            if await self.middleware.call('failover.status') == 'BACKUP':
-                return
-
-        await self.middleware.call("datastore.delete", "system.alert", [])
-
-        for alert in self.alerts:
-            d = alert.__dict__.copy()
-            d["klass"] = alert.instance.config.name
-            d["args"] = alert.instance.args()
-            del d["instance"]
-            del d["mail"]
-            await self.middleware.call("datastore.insert", "system.alert", d)
-
-    @private
-    @job(
-        lock="process_alerts",
-        lock_queue_size=None,  # Must be `None` so that alert operations are not discarded
-        transient=True,
-    )
-    async def oneshot_create(self, job: Any, instance: OneShotAlertClass) -> None:
-        """
-        Creates a one-shot alert from the given alert class `instance`.
-
-        Normal alert creation logic will be applied, so if you create an alert with the same `key` as an already
-        existing alert, no duplicate alert will be created.
-
-        :param instance: an instance of an `AlertClass` subclass that also inherits from `OneShotAlertClass`.
-        """
-
-        if not isinstance(instance, OneShotAlertClass):
-            raise CallError(f"Alert {instance!r} is not a one-shot alert class")
-
-        alert = Alert(instance)
-        alert.source = ""
-        alert.node = self.node
-
-        self.__handle_alert(alert)
-
-        self.alerts = [a for a in self.alerts if a.uuid != alert.uuid] + [alert]
-
-        await self.call2(self.s.alert.send_alerts)
-
-    @api_method(AlertOneshotDeleteArgs, AlertOneshotDeleteResult, private=True)
-    @job(
-        lock="process_alerts",
-        lock_queue_size=None,  # Must be `None` so that alert operations are not discarded
-        transient=True,
-    )
-    async def oneshot_delete(self, job: Any, klass: str | List[str], query: Any = None) -> None:
-        """
-        Deletes one-shot alerts of specified `klass` or klasses, passing `query`
-        to `klass.delete` method.
-
-        It's not an error if no alerts matching delete `query` exist.
-
-        :param klass: either one-shot alert class name (without the `AlertClass` suffix), or list thereof.
-        :param query: `query` that will be passed to `klass.delete` method.
-        """
-
-        klasses: list[str]
-        if isinstance(klass, list):
-            klasses = klass
-        else:
-            klasses = [klass]
-
-        deleted = False
-        for klassname in klasses:
-            try:
-                klass_type = AlertClass.by_name[klassname]
-            except KeyError:
-                raise CallError(f"Invalid alert source: {klassname!r}")
-
-            if not issubclass(klass_type, OneShotAlertClass):
-                raise CallError(f"Alert class {klassname!r} is not a one-shot alert class")
-
-            related_alerts, unrelated_alerts = partition(
-                lambda a: (a.node, a.instance.config.name) == (self.node, klass_type.config.name),
-                self.alerts
-            )
-            left_alerts = await klass_type.delete(related_alerts, query)
-            for deleted_alert in related_alerts:
-                if deleted_alert not in left_alerts:
-                    self.alerts.remove(deleted_alert)
-                    deleted = True
-
-        if deleted:
-            # We need to flush alerts to the database immediately after deleting oneshot alerts.
-            # Some oneshot alerts can only de deleted programmatically (i.e. cloud sync oneshot alerts are deleted
-            # when deleting cloud sync task). If we delete a cloud sync task and then reboot the system abruptly,
-            # the alerts won't be flushed to the database and on next boot an alert for nonexisting cloud sync task
-            # will appear, and it won't be deletable.
-            await self.call2(self.s.alert.flush_alerts)
-
-            await self.call2(self.s.alert.send_alerts)
-
-    @private
-    def alert_source_clear_run(self, name: str) -> None:
-        """
-        Mark the alert source as never ran so that it will be re-run within the next minute.
-        This is useful when you know some alert conditions were just changed.
-
-        :param name: alert source name (without `AlertClass` suffix)
-        """
-        alert_source = self.alert_sources.get(name)
-        if not alert_source:
-            raise CallError(f"Alert source {name!r} not found.", errno.ENOENT)
-
-        self.alert_source_last_run[alert_source.name] = datetime.min
-
-    @private
-    async def product_type(self) -> str:
-        return await self.middleware.call("system.product_type")  # type: ignore[no-any-return]
-
-
-async def _event_system(middleware: Middleware, event_type: str, args: dict[str, Any]) -> None:
-    if SEND_ALERTS_ON_READY:
-        await middleware.call2(middleware.services.alert.send_alerts)
-
-
-async def setup(middleware: Middleware) -> None:
-    await middleware.call2(middleware.services.alertservice.initialize)
-
-    await middleware.call2(middleware.services.alert.initialize)
-
-    middleware.event_subscribe("system.ready", _event_system)
