+++ src/middlewared/middlewared/plugins/alert/alertclasses_config.py
@@ -1,20 +1,9 @@
 from __future__ import annotations
 
-import typing
-
 from middlewared.alert.base import AlertClass
-from middlewared.api import api_method
-from middlewared.api.current import (
-    AlertClassesEntry,
-    AlertClassesUpdate,
-    AlertClassesUpdateArgs,
-    AlertClassesUpdateResult,
-)
-from middlewared.service import ConfigServicePart, GenericConfigService, ValidationErrors
+from middlewared.api.current import AlertClassesEntry, AlertClassesUpdate
+from middlewared.service import ConfigServicePart, ValidationErrors
 import middlewared.sqlalchemy as sa
-
-if typing.TYPE_CHECKING:
-    from middlewared.main import Middleware
 
 
 class AlertClassesModel(sa.Model):
@@ -52,27 +41,3 @@ class AlertClassesConfigServicePart(ConfigServicePart[AlertClassesEntry]):
         await self.middleware.call("datastore.update", self._datastore, old.id, new)
 
         return await self.config()
-
-
-__all__ = ('AlertClassesService',)
-
-
-class AlertClassesService(GenericConfigService[AlertClassesEntry]):
-
-    class Config:
-        datastore = 'system.alertclasses'
-        cli_namespace = 'system.alert.class'
-        entry = AlertClassesEntry
-        role_prefix = 'ALERT'
-        generic = True
-
-    def __init__(self, middleware: Middleware) -> None:
-        super().__init__(middleware)
-        self._svc_part = AlertClassesConfigServicePart(self.context)
-
-    @api_method(AlertClassesUpdateArgs, AlertClassesUpdateResult, check_annotations=True)
-    async def do_update(self, data: AlertClassesUpdate) -> AlertClassesEntry:
-        """
-        Update default Alert settings.
-        """
-        return await self._svc_part.do_update(data)
